requirements.txt
fastapi
uvicorn
google-generativeai
edge-tts
pydantic